Feature/task 14 move logic to services close #14

+++ blog-api/src/main/java/me/academeg/api/service/impl/ImageServiceImpl.java
@@ -1,12 +1,18 @@
 package me.academeg.api.service.impl;
 
 import me.academeg.api.entity.Image;
+import me.academeg.api.exception.entity.ImageNotExistException;
 import me.academeg.api.repository.ImageRepository;
 import me.academeg.api.service.ImageService;
+import me.academeg.api.utils.ImageUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.UUID;
+
+import static me.academeg.api.Constants.AVATAR_PATH;
+import static me.academeg.api.Constants.IMAGE_PATH;
 
 /**
  * ImageServiceImpl
@@ -25,27 +31,31 @@ public class ImageServiceImpl implements ImageService {
     }
 
     @Override
-    public Image add(Image image) {
+    public Image create(MultipartFile file) {
+        Image image = new Image();
+        image.setOriginalPath(ImageUtils.saveImage(IMAGE_PATH, file));
+        image.setThumbnailPath(ImageUtils.compressImage(image.getOriginalPath(), IMAGE_PATH));
         return imageRepository.save(image);
     }
 
     @Override
-    public Image edit(Image image) {
+    public Image update(Image image) {
         return imageRepository.save(image);
     }
 
     @Override
-    public Image getByUuid(UUID uuid) {
-        return imageRepository.findOne(uuid);
+    public Image getById(UUID id) {
+        return imageRepository.findOne(id);
     }
 
     @Override
-    public void delete(Image image) {
+    public void delete(UUID id) {
+        Image image = imageRepository.findOne(id);
+        if (image == null || image.getArticle() == null) {
+            throw new ImageNotExistException();
+        }
+        image.getArticle().getImages().remove(image);
+        ImageUtils.deleteImages(IMAGE_PATH, image.getOriginalPath(), image.getThumbnailPath());
         imageRepository.delete(image);
-    }
-
-    @Override
-    public void delete(UUID uuid) {
-        imageRepository.delete(uuid);
     }
 }
